feat: signup 페이지 interface types에 추가
