+++ src/types/interface.ts
@@ -1,0 +1,7 @@
+// 회원가입 페이지
+export interface UserData {
+  email: string;
+  nickname: string;
+  password: string;
+  confirmPassword: string;
+}
